requirements.txt
dash
bs4
pillow
pandas
dash_bootstrap_components